New features        : Problems solved     : Fixed single precision variants of linear algebra stuff. Bugzilla-Bug-ID     : Additional comments :
git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@231 4b12f95d-9313-0410-afc9-bd26427041f1

+++ kernel/LinearAlgebra/linearalgebra.f90
@@ -279,7 +279,7 @@ CONTAINS
 !<input>
 
   ! Multiplication factor
-  REAL(DP), INTENT(IN) :: sc
+  REAL(SP), INTENT(IN) :: sc
 
 !</input>
